feat:verifica se existe alguma avaliacao vinculada ao aluno

+++ backend/src/use-cases/delete-user-use-case.ts
@@ -9,6 +9,12 @@ export class DeleteUserUseCase {
       throw new Error('Usuário não encontrado');
     }
 
+    // verifica se o usuário é aluno e possui avaliações vinculadas
+    const evaluationsCount = await this.usersRepository.countEvaluationsByStudentId(userIdToDelete);
+    if (evaluationsCount > 0) {
+      throw new Error('Não é possível excluir usuário com avaliações vinculadas');
+    }
+
     await this.usersRepository.delete(userIdToDelete);
   }
 }
